requirements.txt
transformers==4.38.2
einops
accelerate
langchain
bitsandbytes
sentence_transformers
llama_index
torch
llama-index-readers-github
llama-index-embeddings-huggingface
chromadb
llama-index-llms-llama-cpp
llama-index-llms-llama-api
pinecone-client
llama-index-vector-stores-pinecone
pypdf